+++ changelog.md
@@ -4,9 +4,14 @@ All notable changes to `laravel-envcoder` will be documented in this file.
 
 ## Version 1.0.2
 
+### Added
+- env:compare command to show differences between .env and .env.enc
+
 ### Fixed
 - Halt execution on errors from commands
 - Change ask prompt to password to hide user input
+- Allowed use of "" in env values
+- Wrap strings containing spaces with "'s
 
 ## Version 1.0.1